backend/requirements.txt
fastapi
uvicorn[standard]
openai
sentence-transformers
faiss-cpu
sqlalchemy
psycopg2-binary
python-dotenv
PyPDF2
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.1.2
python-docx
pydantic[email]
python-multipart
pydantic-settings
textblob
langdetect